[chore] create issue templates for bugs and features
Created issue templates for community members to report bugs or request features. This should help guide community members on what to include, so issues have all the relevant information for project members to respond effectively.

I tried to make these fairly streamlined while indicating all the information that might be helpful. No suggestion to include an MRP on this project, since people mostly seem to use GameDemo as a reference and would presumably only be reporting bugs that are already present, but I'd probably add it to templates for other projects.