feat: treat CS9057 as error
CS9057 is a compiler warning issued when the game project is using
an earlier version of the .NET compiler than was used to compile the
Introspection generator. Such a mismatch can have serious,
hard-to-diagnose downstream consequences for users. This change
treats CS9057 as an error, to make those issues easier to diagnose and
fix.